haproxy: make maxconn parameter configurable (bsc#1068962)
This can be used to limit or raise the number of connections
balanced to the backends. by default we have a global 4096 limit
over all services, and an individual per-service 1000 concurrent
connection limit. for databases typically 1000 is not enough, so
we need to be able to raise it.

(cherry picked from commit 7b83f135826e1cacb1fe7b5e877a26190ff75b58)