pacemaker: Change default for clone_stateless_services to false
As a reminder, with this option set to false, pacemaker will not manage
stateless services such as OpenStack API services -- instead, systemd
will restart them on crash.

This works rather well, and since we never went to the point where we
had some OCF resource agent for such services, and only use a systemd
resource agent, the benefit of pacemaker is too limited. Reducing
complexity, and converging with the Ardana deployment is therefore more
interesting here.