pacemaker: Do not allow new deployments with DRBD
DRBD is only useful for PostgreSQL HA and unclustered RabbitMQ HA, which
are both features we want to disallow for new deployments. So DRBD has
no real use in new deployments.

For new deployments, we raise an error. For existing deployments (that
are upgraded from a previous release), we let things go happily.