pacemaker: add failure nodes to sync fail message (bsc#1083721)
The current "Some cluster nodes didn't set #{mark}!" isn't a very
usful message. And is a pretty common L3 bug with this simple message.

This is not very helpful for a bug report and it is a distraction from
the actual issue that happened elsewhere, as the customer could have been
looking at other nodes' logs instead of opening an SR right away.

This patch cleans up the logging to include not only the nodes that failed
but also adds the cluster name to each relevant log line. So it's clear
from the logs what pacemaker cluster the message is in reference to.
Which is useful for deployments with more then 1 HA cluster.

Adam already had a PR (202) to solve this, but unfortuantly the
underlying layout changed before it got a chance to merge. So
this patch has become a merge of both our patches to finally solve this
issue.

(cherry picked from commit 90bf282f955ed7ceb34afe2d65f189b6b1ceb1a9)