haproxy: active-active mode, just one VIP
Deploys the haproxy in active-active mode. All nodes in cluster will
have haproxy. Still only one VIP will be active. This is still
beneficial because the 'slave' haproxy instances will be monitoring
the services and a migration of the VIP to one of those nodes will
be faster.

Allows the migration from previous mode active-pasive