DRBD: Fix DRBD resources setup  on reinstall node
When a node is reinstalled the DRBD resources don't are configured
because of the node attribute drbd.rsc.<resource_name>.configured
is marked as true, and the setup is skiped.

This solution store a counter instead of true and false. In each
reinstalation this counter is increased by crowbar. So checking
if the current reinstalation number is equal to the configured
stored in the node attributes it can skip when it's necesary
and setup the DRBD resources when it's needed.