Mark the nodes intended for DRBD setup with proper attributes
Such attributes are used for location constraints that should
make sure DRBD-based resources are only run on nodes with DRBD config.

During the upgrade, cluster can have extra node that is not configured
for DRBD, so that's why we have to distinguish between nodes in DRBD cluser.

Make sure new attributes for existing DRBD are set already during migration,
so that we do not have chef run on existing (unmodified) cluster without them.
It could cause the creation of new constraints and stopping DRBD-based services.