upgrade: Add a healthcheck for pacemaker nodes state
It could end badly if we proceed with the upgrade while some node
is in maintenance state (stopping corosync on such node might lead
to pacemaker's confusion about the services running there).

So let's add a specific check that is executed before the upgrade
and which does not allow user to upgrade such setup.