Drop "clone_stateless_services" attribute
With 5.0 we switch the default of "clone_stateless_services" to false to
have stateless services be managed (restart on failure) by systemd
instead of creating pacemaker resources. When upgrading, existing clouds
will also switch to this configuraton. This patch removes the attribute
for the proposal and makes the systemd based approach this only deploy
option