crowbar-pacemaker: bump sync mark timeout (SCRD-8179)
We are currently seeing sync mark timeouts in Monasca enabled
HA clouds. The reason for these turns out to be the other
controller being a fairly consistent 70s behind the cluster
founder. This causes the sync marks to time out at a timeout
value of 60s. Further investigation didn't reveal any single
culprit (various things added a couple of seconds here and
there and in the end that just summed up to 70s).

This commit bumps the sync mark timeout to 120s to give us a
reasonable amount of breathing space. For we might be able to
shave off a few seconds here and there to get back below the
60s threshold, but that would be very vulnerable to sporadic
delays and/or additional small delays introduced by running
extra code on the cluster founder.