improve galera HA setup (bsc#1122875)
It turns out that haproxy has a known issue with concurrent connections
when using the sticky-table with size 1
Also, there is an scenario using galera that keeps the connections open
to a backend that's declared down. It happens when the health check
fails but the backend still keeps the port open.
This patch fixes both issues increasing the size table to 1000 and
adding an option to shutdown the connections once a backend is marked
down.

Signed-off-by: aojeagarcia <aojeagarcia@suse.com>
(cherry picked from commit 0afb9d334bf5f195b065666e099265f4e5db7b05)