pacemaker: don't check cluster members on founder (SCRD-8462)
When applying the pacemaker barclamp, the founder node first
(re)configures the cluster services on the local host and then
runs a loop waiting for two things to happen:
 - the cluster to be online
 - cluster membership to reflect the nodes that were configured
 (or updated) in the barclamp

This cluster check is not critical - it just results in a warning
being logged if the expected cluster state is not reached in the
configured amount of time (60 seconds).

While the founder node is waiting on the cluster to converge, the
non-founder nodes are stuck in the chef compile phase, also running
a time-boxed while-loop, waiting for the founder node to complete
the setup. Given that this happens in the compile phase, these nodes
haven't yet had a chance to apply the cluster configuration locally,
which means that the founder cluster check will always take the full
60/120 seconds to run and eventually fail.

The same cluster check that is executed on the founder node is
also executed on the non-founder nodes during convergence phase.
However, the difference is that the cluster configuration is actually
created/updated before this happens.

This commit changes the founder cluster check logic so that the
cluster membership part is only run on the non-founder nodes.
This way, the convergence loop run on the founder node has a chance
to complete ahead of the 60 seconds timeout, therefore reducing
the runtime and decreasing the chance that the non-founder nodes
will timeout while waiting for the founder node to complete cluster
setup.

This is required for the Crowbar ECP deployments (SCRD-8462), which
otherwise fail too often because the non-founder nodes run into
timeouts. The other solution would be to increase the timeout
period for the compile phase loop, but that would only hide the real
issue.

(cherry picked from commit 1c2c80f526d98cffcfd50e888725210661e74f07)