pacemaker: wait more for founder if SBD is configured (SCRD-8462)
When SBD is configured for STONITH, the founder node needs to
wait for the cluster to come online twice: one time after the
initial corosync setup and a second time after configuring
the SBD service. Because of this, non-founder nodes also need
to give the founder node more time to complete its setup before
they time out.