Drop g-haproxy location before group deletion (bsc#1156914)
When container (group) resource is deleted, it's children are not
and all constraints assigned to the container are re-assigned to one
of the children. In g-haproxy case this leads to incorrect situation
where l-g-haproxy-controller constraint is assigned to
vip-admin-cluster-data.
Removing the constraint before the group should avoid such situations.

(cherry picked from commit 8540c40fbbd1f67593b467b94aa11cd6f43fcfdd)