add ssl termination on haproxy (bsc#1149535)
If the requests are passed in tcp mode, the source ip gets replaced with
the one of the node where haproxy lives, and there is no way to get the
original ip on the services side.

Add an ability to terminate ssl on haproxy. Using `terminate_ssl` and
`pemfile` will make haproxy switch to http mode and make plain https
requests to the service.